requirements.txt
altair==6.0.0
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cssselect2==0.9.0
distro==1.9.0
docstring_parser==0.18.0
et_xmlfile==2.0.0
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.1
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.56.0
tenacity==9.1.4
tinycss2==1.5.1
tinyhtml5==2.1.0
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
